Separate commands for feedback
For the Gather-Files function, separating the Get-ChildItem from the rest of the pipeline allows us to provide a bit more feedback to the user when running the command about how many files it needs to process for the report. This takes a minor performance hit compared to some testing of doing a Foreach-object loop with write-progress.